Added EventsPage.js
