+++ src/App.js
@@ -4,12 +4,14 @@ import { Navigation } from './Components/Navigation/Navigation';
 import HomePage from './Components/HomePage/HomePage';
 import Footer from './Components/Footer/Footer';
 import 'swiper/css/swiper.css';
+import EventsPage from './Components/EventsPage/EventsPage';
 
 function App() {
   return (
     <div className="App">
-      <Navigation />
-      <HomePage />
+      {/* <Navigation />
+      <HomePage /> */}
+      <EventsPage />
       <Footer />
     </div>
   );
