+++ src/Components/EventsPage/Calendar/Calendar.css
@@ -3,6 +3,7 @@
   width: 100%;
   height: 100%;
   padding: 2px;
+
   background: linear-gradient(to bottom, #00bcf2, #b4a0ff);
 }
 
@@ -108,4 +109,8 @@
   .calendar-container {
     display: none;
   }
+}
+
+.calendar-container {
+  margin: 2%;
 }
