+++ src/Components/HomePage/HomePage.js
@@ -1,11 +1,11 @@
 /* eslint-disable arrow-body-style */
-import React from "react";
-import "./HomePage.css";
-import Featured from "./Featured/Featured";
-import LimeLightZone from "./LimeLightZone/LimeLightZone";
-import AnnouncementBar from "./Announcementbar/AnnouncementBar";
-import Appgrid from "./Appgrid/Appgrid";
-import ClubsInfo from "./ClubsInfo/ClubSlide";
+import React from 'react';
+import './HomePage.css';
+import Featured from './Featured/Featured';
+import LimeLightZone from './LimeLightZone/LimeLightZone';
+import AnnouncementBar from './Announcementbar/AnnouncementBar';
+import Appgrid from './Appgrid/Appgrid';
+import ClubsInfo from './ClubsInfo/ClubSlide';
 
 const HomePage = () => {
   return (
